move id to after if check
