+++ src/packages/shared-utils/index.ts
@@ -1,12 +1,2 @@
-import { CMS_ROLES, CognitoUserAttributes } from '../shared-types'
-
-export const isCmsUser = (user: CognitoUserAttributes) => {
-  const userRoles = user["custom:cms-roles"];
-
-  for (const cmsRole of CMS_ROLES) {
-    if (userRoles.includes(cmsRole)) {
-      return true;
-    }
-  }
-  return false;
-};
+export * from "./is-cms-user";
+export * from "./s3-url-parser";
